+++ apps/frontend/src/components/layout/Footer.tsx
@@ -43,7 +43,7 @@ const Footer: React.FC = () => {
       <Center>
         <Text size="xs" c="dimmed" ta="center">
           This website is a fan-made project and is not affiliated with or endorsed by Blizzard Entertainment, Warcraft Logs, or Raider.IO.<br />
-          Data provided by Warcraft Logs and Raider.IO.<br />
+          Data provided by Blizzard Entertainment, Warcraft Logs and Raider.IO.<br />
           World of Warcraft® is a registered trademark of Blizzard Entertainment, Inc.<br />
           <Anchor component={Link} to="/privacy-policy" size="xs" c="dimmed">
             Privacy Policy